refactor: reorganize ascent logic into `annapurna-logic` crate
The example uses of ascent have been removed, and the ascent logic is
moved into a second crate. The `annapurna-cli` crate is now focused on
reading parameters/configuration so that the logic crate can focus
solely on the logic.
