+++ Cargo.toml
@@ -2,6 +2,22 @@
 members = ["crates/*"]
 resolver = "2"
 
+[workspace.package]
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+homepage = "https://github.com/justinrubek/annapurna"
+repository = "https://github.com/justinrubek/annapurna"
+
+[workspace.dependencies]
+axum = "0.6.4"
+anyhow = "1.0.68"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1.0.87"
+tokio = { version = "1", features = ["full"] }
+thiserror = "1.0.38"
+tracing = "0.1.37"
+
 [profile.release]
 # opt-level = 2 # fast and small wasm
