+++ crates/cli/src/main.rs
@@ -1,38 +1,12 @@
+use annapurna_logic::{
+    recipe,
+    types::{Ingredient, Recipe},
+};
 use clap::Parser;
 use std::collections::HashMap;
 
-mod aggregators;
-mod example;
-
-use example::run as example_run;
-mod named;
-use named::run as named_run;
-
-mod recipe;
-use recipe::RecipeManager;
-
-pub mod program;
-use program::AscentProgram;
-use types::{Ingredient, Recipe};
-
 pub mod commands;
 use commands::{BasicCommands, Commands};
-pub mod types;
-
-#[cfg(test)]
-mod tests;
-
-fn recipe(recipes: Vec<Recipe>, has_ingredients: Vec<Ingredient>) {
-    // convert to String
-    let manager = RecipeManager::new(has_ingredients, recipes);
-
-    let res = manager.process();
-
-    let can_make = res.can_make;
-    let missing = res.missing;
-    println!("Can make: {can_make:?}");
-    println!("Missing: {missing:?}");
-}
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     tracing_subscriber::fmt::init();
@@ -60,14 +34,4 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     }
 
     Ok(())
-}
-
-#[allow(dead_code)]
-fn example() {
-    let edges = vec![(1, 2), (2, 3)];
-    let connected = example_run(edges.clone());
-    println!("Connected: {connected:?}");
-
-    let connected = named_run(edges);
-    println!("Connected: {connected:?}");
 }
