+++ crates/logic/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "annapurna-logic"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+ascent = "0.4.0"
+anyhow.workspace = true
+serde.workspace = true
+# serde_json = "1.0.87"
+thiserror.workspace = true
+tracing.workspace = true
