+++ crates/logic/src/error.rs
@@ -1,0 +1,7 @@
+#[derive(thiserror::Error, Debug)]
+pub enum Error {
+    #[error("todo")]
+    Todo,
+}
+
+pub type Result<T> = std::result::Result<T, Error>;
