+++ crates/logic/src/tests.rs
